# swagger.yaml
{
  "swagger": "2.0",
  "info": {
    "title": "api/api.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/alternate-logo": {
      "get": {
        "operationId": "AlternateLogo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/serverAlternateLogoOutput"
            }
          }
        },
        "tags": [
          "Server"
        ]
      }
    },
    "/api/crew": {
      "get": {
        "operationId": "Crew",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/crewCrew"
            }
          }
        },
        "tags": [
          "Server"
        ]
      }
    },
    "/api/dashboard": {
      "get": {
        "operationId": "Dashboard",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/dashboardEntries"
            }
          }
        },
        "tags": [
          "Server"
        ]
      }
    },
    "/api/hackz": {
      "get": {
        "operationId": "Hackz",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/dashboardEntries"
            }
          }
        },
        "tags": [
          "Server"
        ]
      }
    },
    "/api/kryptos/decrypt": {
      "post": {
        "operationId": "KryptosDecrypt",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/serverKryptosOutput"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/serverKryptosInput"
            }
          }
        ],
        "tags": [
          "Server"
        ]
      }
    },
    "/api/kryptos/encrypt": {
      "post": {
        "operationId": "KryptosEncrypt",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/serverKryptosOutput"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/serverKryptosInput"
            }
          }
        ],
        "tags": [
          "Server"
        ]
      }
    },
    "/api/moijaime": {
      "get": {
        "operationId": "Moijaime",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/serverMoijaimeOutput"
            }
          }
        },
        "tags": [
          "Server"
        ]
      }
    },
    "/api/numberinfo/{number}": {
      "get": {
        "operationId": "Numberinfo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/serverNumberinfoOutput"
            }
          }
        },
        "parameters": [
          {
            "name": "number",
            "in": "path",
            "required": true,
            "type": "number",
            "format": "float"
          }
        ],
        "tags": [
          "Server"
        ]
      }
    },
    "/api/ping": {
      "get": {
        "operationId": "Ping",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/serverPong"
            }
          }
        },
        "tags": [
          "Server"
        ]
      }
    },
    "/api/recettator": {
      "get": {
        "operationId": "Recettator",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/serverRecettatorOutput"
            }
          }
        },
        "parameters": [
          {
            "name": "seed",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "main_ingredients",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "uint64"
          },
          {
            "name": "secondary_ingredients",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "uint64"
          },
          {
            "name": "steps",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "Server"
        ]
      }
    },
    "/api/soundcloud/me": {
      "get": {
        "operationId": "SoundcloudMe",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/soundcloudUser"
            }
          }
        },
        "tags": [
          "Server"
        ]
      }
    },
    "/api/soundcloud/playlists": {
      "get": {
        "operationId": "SoundcloudPlaylists",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/soundcloudPlaylists"
            }
          }
        },
        "tags": [
          "Server"
        ]
      }
    },
    "/api/soundcloud/playlists/{playlist_id}": {
      "get": {
        "operationId": "SoundcloudPlaylist",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/soundcloudPlaylist"
            }
          }
        },
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "Server"
        ]
      }
    },
    "/api/soundcloud/tracks": {
      "get": {
        "operationId": "SoundcloudTracks",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/soundcloudTracks"
            }
          }
        },
        "tags": [
          "Server"
        ]
      }
    },
    "/api/soundcloud/tracks/{track_id}": {
      "get": {
        "operationId": "SoundcloudTrack",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/soundcloudTrack"
            }
          }
        },
        "parameters": [
          {
            "name": "track_id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "Server"
        ]
      }
    },
    "/api/spreadshirt/all": {
      "get": {
        "operationId": "SpreadshirtAll",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/serverSpreadshirtAllOutput"
            }
          }
        },
        "tags": [
          "Server"
        ]
      }
    },
    "/api/spreadshirt/random": {
      "get": {
        "operationId": "SpreadshirtRandom",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/serverSpreadshirtRandomOutput"
            }
          }
        },
        "tags": [
          "Server"
        ]
      }
    },
    "/api/tpyo": {
      "post": {
        "operationId": "TpyoEnocde",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/serverTpyoEnocdeOuptut"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/serverTpyoEnocdeIpunt"
            }
          }
        ],
        "tags": [
          "Server"
        ]
      }
    },
    "/api/wotd": {
      "get": {
        "operationId": "Wotd",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/serverWotdOutput"
            }
          }
        },
        "tags": [
          "Server"
        ]
      }
    }
  },
  "definitions": {
    "EntryKind": {
      "type": "string",
      "enum": [
        "Undefined",
        "Hack",
        "Track",
        "Merch"
      ],
      "default": "Undefined"
    },
    "crewCrew": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "website": {
          "type": "string"
        },
        "members": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/crewPerson"
          }
        },
        "accounts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/crewWebAccount"
          }
        },
        "friends": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/crewFriend"
          }
        }
      }
    },
    "crewFriend": {
      "type": "object",
      "properties": {
        "key": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "logo_url": {
          "type": "string"
        },
        "links": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/crewLink"
          }
        },
        "image_url": {
          "type": "string"
        }
      }
    },
    "crewLink": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "url": {
          "type": "string"
        }
      }
    },
    "crewPerson": {
      "type": "object",
      "properties": {
        "key": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "crewWebAccount": {
      "type": "object",
      "properties": {
        "key": {
          "type": "string"
        },
        "provider": {
          "type": "string"
        },
        "handle": {
          "type": "string"
        },
        "url": {
          "type": "string"
        }
      }
    },
    "dashboardEntries": {
      "type": "object",
      "properties": {
        "entries": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/dashboardEntry"
          }
        }
      }
    },
    "dashboardEntry": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "url": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "image_url": {
          "type": "string"
        },
        "is_external": {
          "type": "boolean",
          "format": "boolean"
        },
        "kind": {
          "$ref": "#/definitions/EntryKind"
        }
      }
    },
    "serverAlternateLogoOutput": {
      "type": "object",
      "properties": {
        "path": {
          "type": "string"
        }
      }
    },
    "serverKryptosInput": {
      "type": "object",
      "properties": {
        "from": {
          "type": "string"
        }
      }
    },
    "serverKryptosOutput": {
      "type": "object",
      "properties": {
        "to": {
          "type": "string"
        }
      }
    },
    "serverMoijaimeOutput": {
      "type": "object",
      "properties": {
        "kiffs": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "serverNumberinfoOutput": {
      "type": "object",
      "properties": {
        "facts": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        }
      }
    },
    "serverPong": {
      "type": "object",
      "properties": {
        "pong": {
          "type": "string"
        }
      }
    },
    "serverRecettatorIngredient": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "quantity": {
          "type": "string"
        },
        "method": {
          "type": "string"
        },
        "gender": {
          "type": "string"
        },
        "multiple": {
          "type": "boolean",
          "format": "boolean"
        },
        "kind": {
          "type": "string"
        },
        "name_and_quantity": {
          "type": "string"
        }
      }
    },
    "serverRecettatorOutput": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "people": {
          "type": "string",
          "format": "uint64"
        },
        "markdown": {
          "type": "string"
        },
        "steps": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "main_ingredients": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/serverRecettatorIngredient"
          }
        },
        "secondary_ingredients": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/serverRecettatorIngredient"
          }
        },
        "seed": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "serverSpreadshirtAllOutput": {
      "type": "object"
    },
    "serverSpreadshirtRandomOutput": {
      "type": "object"
    },
    "serverTpyoEnocdeIpunt": {
      "type": "object",
      "properties": {
        "form": {
          "type": "string"
        }
      },
      "title": "tpyo megasess"
    },
    "serverTpyoEnocdeOuptut": {
      "type": "object",
      "properties": {
        "to": {
          "type": "string"
        }
      }
    },
    "serverWotdOutput": {
      "type": "object",
      "properties": {
        "word": {
          "type": "string"
        }
      }
    },
    "soundcloudPlaylist": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uint64"
        },
        "created_at": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "sharing": {
          "type": "string"
        },
        "embeddable_by": {
          "type": "string"
        },
        "purchase_url": {
          "type": "string"
        },
        "artwork_url": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "duration": {
          "type": "string",
          "format": "uint64"
        },
        "genre": {
          "type": "string"
        },
        "shared_to_count": {
          "type": "string",
          "format": "uint64"
        },
        "tag_list": {
          "type": "string"
        },
        "release_day": {
          "type": "integer",
          "format": "int64",
          "title": "uint64 release = 13;"
        },
        "release_month": {
          "type": "integer",
          "format": "int64"
        },
        "release_year": {
          "type": "integer",
          "format": "int64"
        },
        "streamable": {
          "type": "boolean",
          "format": "boolean"
        },
        "downloadable": {
          "type": "boolean",
          "format": "boolean"
        },
        "ean": {
          "type": "string"
        },
        "playlist_type": {
          "type": "string"
        },
        "tracks": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/soundcloudTrack"
          }
        },
        "uri": {
          "type": "string"
        },
        "label": {
          "$ref": "#/definitions/soundcloudUser"
        },
        "label_id": {
          "type": "string",
          "format": "uint64"
        },
        "label_name": {
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/soundcloudUser"
        },
        "user_id": {
          "type": "string",
          "format": "uint64"
        },
        "permalink": {
          "type": "string"
        },
        "permalink_url": {
          "type": "string"
        }
      }
    },
    "soundcloudPlaylists": {
      "type": "object",
      "properties": {
        "playlists": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/soundcloudPlaylist"
          }
        }
      }
    },
    "soundcloudTrack": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uint64"
        },
        "CreatedAt": {
          "type": "string"
        },
        "Title": {
          "type": "string"
        },
        "Sharing": {
          "type": "string"
        },
        "EmbeddableBy": {
          "type": "string"
        },
        "PurchaseUrl": {
          "type": "string"
        },
        "ArtworkUrl": {
          "type": "string"
        },
        "Description": {
          "type": "string"
        },
        "Duration": {
          "type": "string",
          "format": "uint64"
        },
        "Genre": {
          "type": "string"
        },
        "SharedToCount": {
          "type": "string",
          "format": "uint64"
        },
        "TagList": {
          "type": "string"
        },
        "ReleaseDay": {
          "type": "integer",
          "format": "int64",
          "title": "uint64 Release = 13;"
        },
        "ReleaseMonth": {
          "type": "integer",
          "format": "int64"
        },
        "ReleaseYear": {
          "type": "integer",
          "format": "int64"
        },
        "Streamable": {
          "type": "boolean",
          "format": "boolean"
        },
        "Downloadable": {
          "type": "boolean",
          "format": "boolean"
        },
        "State": {
          "type": "string"
        },
        "License": {
          "type": "string"
        },
        "TrackType": {
          "type": "string"
        },
        "WaveformUrl": {
          "type": "string"
        },
        "DownloadUrl": {
          "type": "string"
        },
        "StreamUrl": {
          "type": "string"
        },
        "VideoUrl": {
          "type": "string"
        },
        "Bpm": {
          "type": "number",
          "format": "float"
        },
        "Commentable": {
          "type": "boolean",
          "format": "boolean"
        },
        "ISRC": {
          "type": "string"
        },
        "KeySignature": {
          "type": "string"
        },
        "CommentCount": {
          "type": "string",
          "format": "uint64"
        },
        "DownloadCount": {
          "type": "string",
          "format": "uint64"
        },
        "PlaybackCount": {
          "type": "string",
          "format": "uint64"
        },
        "FavoritingsCount": {
          "type": "string",
          "format": "uint64"
        },
        "OriginalFormat": {
          "type": "string"
        },
        "OriginalContentSize": {
          "type": "string",
          "format": "uint64"
        },
        "AssetData": {
          "type": "string",
          "format": "byte",
          "title": "App created_with = 36;"
        },
        "ArtworkData": {
          "type": "string",
          "format": "byte"
        },
        "UserFavorite": {
          "type": "boolean",
          "format": "boolean"
        },
        "uri": {
          "type": "string"
        },
        "label": {
          "$ref": "#/definitions/soundcloudUser"
        },
        "label_id": {
          "type": "string",
          "format": "uint64"
        },
        "label_name": {
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/soundcloudUser"
        },
        "user_id": {
          "type": "string",
          "format": "uint64"
        },
        "permalink": {
          "type": "string"
        },
        "permalink_url": {
          "type": "string"
        }
      }
    },
    "soundcloudTracks": {
      "type": "object",
      "properties": {
        "tracks": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/soundcloudTrack"
          }
        }
      }
    },
    "soundcloudUser": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uint64"
        },
        "username": {
          "type": "string"
        },
        "avatar_url": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "full_name": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "discogs_name": {
          "type": "string"
        },
        "myspace_name": {
          "type": "string"
        },
        "website": {
          "type": "string"
        },
        "website_title": {
          "type": "string"
        },
        "online": {
          "type": "boolean",
          "format": "boolean"
        },
        "track_count": {
          "type": "string",
          "format": "uint64"
        },
        "playlist_count": {
          "type": "string",
          "format": "uint64"
        },
        "followers_count": {
          "type": "string",
          "format": "uint64"
        },
        "followings_count": {
          "type": "string",
          "format": "uint64"
        },
        "public_favorites_count": {
          "type": "string",
          "format": "uint64"
        },
        "avatar_data": {
          "type": "string",
          "format": "byte"
        },
        "uri": {
          "type": "string"
        },
        "permalink": {
          "type": "string"
        },
        "permalink_url": {
          "type": "string"
        }
      }
    }
  }
}